:recycle: Move bookmarkDropdown and bookmarksList inside bookmarks folder
Rename bookmark component to bookmarkDropdown as it was misleading.
Update all imports related to the components
