+++ src/components/bookmarks/bookmarkDropdown/bookmarkDropdown.css
@@ -1,5 +1,5 @@
-@import '../app/variablesV2.css';
-@import '../app/mixins.css';
+@import '../../app/variablesV2.css';
+@import '../../app/mixins.css';
 
 .wrapper {
   box-sizing: border-box;
