+++ src/components/bookmarks/bookmarkDropdown/index.js
@@ -1,8 +1,8 @@
 /* istanbul ignore file */
 import { connect } from 'react-redux';
 import { translate } from 'react-i18next';
-import { bookmarkAdded, bookmarkRemoved } from '../../actions/bookmarks';
-import Bookmark from './bookmark';
+import { bookmarkAdded, bookmarkRemoved } from '../../../actions/bookmarks';
+import BookmarkDropdown from './bookmarkDropdown';
 
 const mapStateToProps = state => ({
   bookmarks: state.bookmarks,
@@ -13,4 +13,4 @@ const mapDispatchToProps = {
   bookmarkRemoved,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(translate()(Bookmark));
+export default connect(mapStateToProps, mapDispatchToProps)(translate()(BookmarkDropdown));
